app/requirement.txt
Flask
torchvision
efficientnet-pytorch
Pillow
lime
scikit-image
matplotlib
numpy
torch